Re formatted Structure
Every module now should have it's own .js .css file, the aim is to make
it so every module can be included in a template titled TEMPLATE_index
with the CSS filed needed included at the top and the JS files needed
included before the footer. We also have a scripts_common.html file with
is a placeholder for files included in all pages regardless and
TEMPLATE.JS and TEMPLATE.CSS files which are solo files for the specific
template pages only.