Add slider arrows to all sliders
This required html add ons to all the snippets.
Also the IOS Slider api is called on the left and right sliders
CSS code is added for each and every arrow
Templates are also updated in the case of Society and Sport pages
All the inline javascript from the mentioned pages was moved rightfully
to su_content.js
IOS Slider was updated to the latest and greatest version.