Changed FTO Election header
This code takes the icon of the page and makes a neat header with it.

This code can be altered to work with other pages as well.

Also consider (for some pages) altering the code in a way that it
replaces a the stretch background of #cover instead of inserting a
centered page icon.